websocket_listener/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
binance==0.3
bottle==0.12.19
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
clickhouse-cityhash==1.0.2.3
clickhouse-driver==0.2.1
dateparser==1.0.0
idna==3.2
lz4==3.1.3
multidict==5.1.0
python-binance==1.0.14
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==4.1.0
urllib3==1.26.6
websockets==10.0
yarl==1.6.3
